version 1.0.7 beta

+++ src/engine/param.h
@@ -15,17 +15,17 @@ constexpr uint8_t EXACT_FLAG = 0;
 constexpr uint8_t ALPHA_FLAG = 1;
 constexpr uint8_t BETA_FLAG = 2;
 
-constexpr int16_t MAX_DEPTH = 255;
-constexpr int16_t TB_DEPTH = 254;
+constexpr int32_t MAX_DEPTH = 255;
+constexpr int32_t TB_DEPTH = 254;
 
-constexpr int32_t MATED_IN(int16_t ply)
+constexpr int32_t MATED_IN(int32_t ply)
 {
-    return -INF + static_cast<int32_t>(ply);
+    return -INF + ply;
 }
 
-constexpr int32_t MATE_IN(int16_t ply)
+constexpr int32_t MATE_IN(int32_t ply)
 {
-    return INF - static_cast<int32_t>(ply);
+    return INF - ply;
 }
 
 } // namespace param
